+++ packages/plugin-js/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @terrazzo/plugin-js
 
+## 2.0.3
+
+### Patch Changes
+
+- [#726](https://github.com/terrazzoapp/terrazzo/pull/726) [`b326bd6`](https://github.com/terrazzoapp/terrazzo/commit/b326bd63b2f701707d0501675bde6d9b091de8fa) Thanks [@drwpow](https://github.com/drwpow)! - Improve token aliases where the primitive or primary has token, but aliases do not. This leads to dropped tokens in all plugins.
+
+- Updated dependencies [[`b326bd6`](https://github.com/terrazzoapp/terrazzo/commit/b326bd63b2f701707d0501675bde6d9b091de8fa)]:
+  - @terrazzo/parser@2.0.3
+  - @terrazzo/cli@2.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+- [#719](https://github.com/terrazzoapp/terrazzo/pull/719) [`2d4014f`](https://github.com/terrazzoapp/terrazzo/commit/2d4014f12de99149d19a2f733a9723daf19c39d2) Thanks [@drwpow](https://github.com/drwpow)! - plugin-css: fix bug with legacy modeSelectors resolving to default alias instead of requested mode
+
+- Updated dependencies [[`2d4014f`](https://github.com/terrazzoapp/terrazzo/commit/2d4014f12de99149d19a2f733a9723daf19c39d2)]:
+  - @terrazzo/cli@2.0.2
+  - @terrazzo/parser@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- [#699](https://github.com/terrazzoapp/terrazzo/pull/699) [`a23b751`](https://github.com/terrazzoapp/terrazzo/commit/a23b7510710943768731859215b92a98355e1c5c) Thanks [@drwpow](https://github.com/drwpow)! - [parser]: fix runtime error where $extensions may not have mode object
+
+- Updated dependencies [[`a23b751`](https://github.com/terrazzoapp/terrazzo/commit/a23b7510710943768731859215b92a98355e1c5c)]:
+  - @terrazzo/parser@2.0.1
+  - @terrazzo/cli@2.0.1
+
 ## 2.0.0
 
 ### Minor Changes