+++ packages/plugin-swift/CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Patch Changes
 
+- [#719](https://github.com/terrazzoapp/terrazzo/pull/719) [`2d4014f`](https://github.com/terrazzoapp/terrazzo/commit/2d4014f12de99149d19a2f733a9723daf19c39d2) Thanks [@drwpow](https://github.com/drwpow)! - plugin-css: fix bug with legacy modeSelectors resolving to default alias instead of requested mode
+
+- Updated dependencies [[`2d4014f`](https://github.com/terrazzoapp/terrazzo/commit/2d4014f12de99149d19a2f733a9723daf19c39d2)]:
+  - @terrazzo/cli@2.0.2
+  - @terrazzo/parser@2.0.2
+  - @terrazzo/token-tools@2.0.2
+
+## 0.3.1
+
+### Patch Changes
+
 - [#553](https://github.com/terrazzoapp/terrazzo/pull/553) [`e63a627`](https://github.com/terrazzoapp/terrazzo/commit/e63a6277f61282fb608744a8348689b16f977076) Thanks [@Sidnioulz](https://github.com/Sidnioulz)! - Add support for the Token Listing format
 
 ## 0.3.0